hospital_management/requirements.txt
Django==4.2.7
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
mssql-django>=1.6
pyodbc>=5.2.0
python-decouple==3.8
django-cors-headers>=4.7.0
django-filter>=25.1
Pillow>=11.3.0
openpyxl>=3.1.5
pandas>=2.3.2
reportlab>=4.4.2
python-dateutil>=2.8.2
pytz>=2023.3
drf-spectacular>=0.27.0
django-health-check>=3.17.0
psutil>=5.9.5
python-json-logger>=2.0.7
